LayoutTests/js/sequence-iterator-protocol-expected.txt

CONSOLE MESSAGE:
CONSOLE MESSAGE: Using an array: 1,2,3,1,2,3
CONSOLE MESSAGE: Using an iterator: 10,11,12,10,11,12
CONSOLE MESSAGE: Using an array but with customized iteration (should be same as iterator): 10,11,12,10,11,12
The second and third results should be the same.

PASS successfullyParsed is true

TEST COMPLETE
